requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
openai==1.12.0
groq==0.4.1
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
fastapi
google-generativeai
groq
motor
openai
pydantic
pydantic-settings
pymongo
python-dotenv
uvicorn
